create new trail game component
